+++ rustls/src/client/tls12.rs
@@ -1,12 +1,12 @@
 use crate::check::{inappropriate_handshake_message, inappropriate_message};
 use crate::common_state::{CommonState, Side, State};
 use crate::conn::ConnectionRandoms;
+use crate::crypto::ring::KeyExchange;
 use crate::crypto::{CryptoProvider, KeyExchangeError};
 use crate::enums::ProtocolVersion;
 use crate::enums::{AlertDescription, ContentType, HandshakeType};
 use crate::error::{Error, InvalidMessage, PeerMisbehaved};
 use crate::hash_hs::HandshakeHash;
-use crate::kx;
 #[cfg(feature = "logging")]
 use crate::log::{debug, trace, warn};
 use crate::msgs::base::{Payload, PayloadU8};
@@ -767,7 +767,7 @@ impl<C: CryptoProvider> State<ClientConnectionData> for ExpectServerDone<C> {
         let ecdh_params =
             tls12::decode_ecdh_params::<ServerECDHParams>(cx.common, &st.server_kx.kx_params)?;
         let named_group = ecdh_params.curve_params.named_group;
-        let kx = match kx::KeyExchange::choose(named_group, &st.config.kx_groups) {
+        let kx = match KeyExchange::choose(named_group, &st.config.kx_groups) {
             Ok(kx) => kx,
             Err(KeyExchangeError::UnsupportedGroup) => {
                 return Err(PeerMisbehaved::SelectedUnofferedKxGroup.into())
